+++ src/asana.js
@@ -10,10 +10,18 @@ var Client   = require("asana").Client;
 // Log
 var chalk = require('chalk');
 
+var color_index = ["red", "green", "yellow", "magenta", "blue",
+                   "bgRed", "bgGreen", "bgCyan", "bgMagenta", "bgBlue"];
+
+var colorize = function(index, string) {
+  return chalk[color_index[index] || "gray"](string);
+};
+
 export class Asana {
   constructor(api_key, options) {
     options = R.merge({
-      to: 'debug.json'
+      to      : 'debug.json',
+      parallel: 10
     }, options);
     options.to   = path.resolve(process.cwd(), options.to);
 
@@ -26,11 +34,19 @@ export class Asana {
   }
 
   run(workspace_name, projects) {
-    utils.log("Get tasks from projects", chalk.green(projects.join(chalk.white(', '))));
+    utils.log("Get tasks from projects", chalk.cyan(projects.join(chalk.white(', '))));
     return BPromise.coroutine(function* () {
-      var workspace = yield this._workspaceByName(workspace_name);
+      var workspace      = yield this._workspaceByName(workspace_name);
       workspace.projects = yield this._projectsByWorkspace(workspace, projects);
-      workspace.tasks    = yield this._tasksAllProjects(projects, workspace);
+
+      var tasks       = yield this._tasksAllProjects(projects, workspace);
+      workspace.tasks = tasks[0];
+      utils.log("Save", chalk.green(workspace.tasks.length), "tasks");
+
+      if (!R.isEmpty(tasks[1])) {
+        workspace.errors = tasks[1] || [];
+        utils.log("with", chalk.green(workspace.errors.length), "erros");
+      }
 
       var json = JSON.stringify(workspace, null, 2);
       this.saveFile(json);
@@ -46,53 +62,118 @@ export class Asana {
       });
   }
 
-  _taskById(task) {
-    utils.log("      Find task by id", chalk.green(task.id), "-", task.name);
-
+  _parallelTasks(elems, padding) {
     return BPromise.coroutine(function* () {
-      task = yield this.client.tasks.findById(task.id);
-      var attachments = [];
-      // var attachments = yield this._attachmentsByTaskId(task.id);
-      var stories     = yield this._storiesByTaskId(task.id);
-      var subtasks    = yield this._subtasksByTaskID(task.id);
+      var success = [];
+      var errors  = [];
+      var retry_after = 0;
 
-      task.attachments = attachments;
-      task.stories     = this._storiesClean(stories);
-      task.subtasks    = subtasks;
+      if (elems.length > 0) {
+        utils.log(padding + "Get content of", chalk.cyan(elems.length), "tasks.\n");
 
-      return task;
+        var parallel = this.options.parallel;
+        var start = 0;
+
+        while (start < elems.length) {
+          var end = parallel + start;
+          var multiple_promisses = [];
+
+          for (var acc = start, i = 0; acc < end; acc++, i++) {
+            var task = elems[acc];
+            multiple_promisses.push(this._taskById(i, task, padding, acc));
+          }
+
+          yield BPromise.settle(multiple_promisses).then((results) => {
+            for (var j in results) {
+              var result = results[j];
+              var elm_i  = Number(j) + start;
+              var elm    = elems[elm_i];
+
+              if (result.isFulfilled()) {
+                success.push(result.value());
+              } else if (result.isRejected()) {
+                retry_after = Math.max(retry_after, result.reason().retryAfterSeconds) || retry_after;
+                elm.error = result;
+                // console.log('result:', result);
+                errors.push(elm);
+                // result.reason();
+              }
+            }
+          });
+
+          start = end;
+        }
+      }
+      if (errors.length > 0) {
+        // console.log(JSON.stringify(errors, null, 2));
+        retry_after += 1;
+        utils.log(padding + "Retry", chalk.cyan(errors.length), "tasks in", retry_after, "seconds.\n");
+        yield BPromise.delay(retry_after * 1000);
+
+        var retry = yield this._parallelTasks(errors, "      ");
+        success.concat(retry[0]);
+        errors = retry[1];
+      }
+
+      return [success, errors];
+    }.bind(this))();
+  }
+
+  _taskById(i, task, padding, acc) {
+    return BPromise.coroutine(function* () {
+      if (task && task.id) {
+        var tag = padding + colorize(i, "@" + acc);
+        utils.log(tag, colorize(i, task.id + ":" ), "Find data of task", chalk.gray(task.name));
+        padding += "  "; // increment padding to chields
+
+        task = yield this.client.tasks.findById(task.id);
+        // var attachments = yield this._attachmentsByTaskId(task.id);
+        // task.attachments = attachments;
+
+        var stories   = yield this._storiesByTaskId(i, task.id, padding, i)
+          .error((err) => {
+            utils.log(padding, " Error to get stories of task", colorize(i, task.id ), err);
+            utils.log(padding, err);
+          });
+        task.stories  = this._storiesClean(stories);
+        task.subtasks = yield this._subtasksByTaskID(i, task.id, padding, i)
+          .error((err) => {
+            utils.log(padding, " Error to get subtasks of task", colorize(i, task.id ), err);
+            utils.log(padding, err);
+          });
+
+        return task;
+      }
     }.bind(this))();
   }
 
   _tasksByProject(project) {
-    utils.log("    Find tasks in", chalk.green(project.id), project.name);
+    utils.log("    " + chalk.cyan(project.id) + ": Find tasks to project", project.name);
     return BPromise.coroutine(function* () {
       var tasks = yield this.client.tasks
         .findByProject(project.id)
         .then((result) => result.data ? result.data : []);
 
-      for (var ix = 0; ix < tasks.length; ix++) {
-        tasks[ix] = yield this._taskById(tasks[ix]);
-      }
-
-      return tasks;
+      return this._parallelTasks(tasks, "      ");
     }.bind(this))();
   }
 
   _tasksAllProjects(projects, workspace) {
     return BPromise.coroutine(function* () {
-      var all_tasks = [];
+      var tasks  = [];
+      var errors = [];
       for (var i = 0; projects.length > i; i++) {
         var filter    = R.find(R.propEq('name', projects[i]));
         var project   = filter(workspace.projects);
 
         if (!R.isEmpty(project) && project.id) {
-          var tasks = yield this._tasksByProject(project);
+          var result = yield this._tasksByProject(project);
 
-          all_tasks = all_tasks.concat(tasks);
+          tasks  = tasks.concat(result[0]);
+          errors = errors.concat(result[1]);
         }
       }
-      return all_tasks;
+      return [tasks, errors];
     }.bind(this))();
   }
 
@@ -115,23 +196,28 @@ export class Asana {
     return this.client.attachments.findById(attachment.id);
   }
 
-  _subtasksByTaskID(id) {
-    utils.log("        Find subtasks from task", chalk.green(id));
+  _subtasksByTaskID(task_index, id, padding, acc) {
+    var tag = padding + colorize(task_index, "@" + acc);
+    utils.log(tag, "Find subtasks from task", colorize(task_index, id));
+    padding += "  "; // increment padding to chields
+
     return BPromise.coroutine(function* () {
       var subtasks = yield this.client.tasks
         .subtasks(id)
         .then((result) => result.data ? result.data : []);
 
       for (var i = 0; subtasks.length > i; i++) {
-        subtasks[i] = yield this._taskById(subtasks[i]);
+        subtasks[i] = yield this._taskById(task_index, subtasks[i], padding, i);
       }
 
       return subtasks;
+      // return this._parallelTasks(subtasks, padding);
     }.bind(this))();
   }
 
-  _storiesByTaskId(id) {
-    utils.log("        Find stories by task id", chalk.green(id));
+  _storiesByTaskId(task_index, id, padding, acc) {
+    var tag = padding + colorize(task_index, "@" + acc);
+    utils.log(tag, "Find stories by task id", colorize(task_index, id));
     return this.client.stories
       .findByTask(id)
       .then((result) => result.data ? result.data : []);